cks-lab-generator/pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "cks-lab-generator"
version = "1.0.0"
description = "Automated CKS Lab Exercises Generator - Generate complete Kubernetes security lab exercises from CKS guide"
readme = "README.md"
requires-python = ">=3.10"
authors = [
    {name = "CKS Lab Generator Team"}
]
keywords = ["kubernetes", "cks", "security", "lab", "generator", "certification"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "Intended Audience :: Education",
    "Topic :: Education :: Testing",
    "Topic :: Software Development :: Code Generators",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]

dependencies = [
    "markdown-it-py>=3.0.0",
    "PyYAML>=6.0.1",
    "click>=8.1.7",
    "Jinja2>=3.1.3",
    "python-dateutil>=2.8.2",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.3",
    "pytest-cov>=4.1.0",
    "pytest-mock>=3.12.0",
    "black>=23.12.1",
    "ruff>=0.1.9",
    "mypy>=1.8.0",
    "types-PyYAML>=6.0.12",
]

[project.scripts]
cks-lab-gen = "cks_lab_generator.cli:main"

[tool.setuptools.packages.find]
where = ["src"]

[tool.black]
line-length = 100
target-version = ['py310', 'py311', 'py312']

[tool.ruff]
line-length = 100
target-version = "py310"
select = ["E", "F", "I", "N", "W", "UP"]

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
ignore_missing_imports = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "-v --cov=src/cks_lab_generator --cov-report=term-missing --cov-report=html"
